requirements.txt
asgiref==3.2.10
backports.zoneinfo==0.2.1
Django==3.1
django-pdfkit==0.3.1
django-tinymce==3.5.0
pdfkit==0.6.0
pytz==2022.6
sqlparse==0.4.3
XlsxWriter==3.0.3
